Prevents osu!.db reader from going into an infinite loop on bad osu!.db.
If it reads goes past the end of the stream with seekg, then iosstream doesn't actually set failbit, badbit, or eofbit. Adding a .peek() will cause those bits to be set. However, those bits are only good if we check them, so after each field in osu!.db is read then fail the init if it detects a bad bit.